requirements.txt
fastapi==0.128.3
uvicorn==0.40.0
pillow==12.1.0
